--- a/option-t.es6.d.ts
+++ b/option-t.es6.d.ts
@@ -26,7 +26,8 @@
  *  The Option/Maybe type interface whose APIs are inspired
  *  by Rust's `std::option::Option<T>`.
  */
-export interface Option<T> {
+export type Option<T> = Some<T> | None<T>;
+interface OptionMethods<T> {
 
     /**
      *  Return whether the self is `Some<T>` or not.
@@ -184,7 +185,7 @@ export interface Option<T> {
  */
 export abstract class OptionBase {}
 
-export class Some<T> extends OptionBase implements Option<T> {
+export class Some<T> extends OptionBase implements OptionMethods<T> {
     constructor(val: T);
     isSome: this is Some<T>;
     isNone: this is None<T>;
@@ -204,7 +205,7 @@ export class Some<T> extends OptionBase implements Option<T> {
     drop(): void;
 }
 
-export class None<T> extends OptionBase implements Option<T> {
+export class None<T> extends OptionBase implements OptionMethods<T> {
     constructor();
     isSome: this is Some<T>;
     isNone: this is None<T>;