--- a/tsconfig.json
+++ b/tsconfig.json
@@ -1,5 +1,6 @@
 {
     "compilerOptions": {
+        "diagnostics": true,
         "noImplicitAny": true,
         "removeComments": false,
         "preserveConstEnums": true,