## Motivation

In this case, we'd like to treat `v` as `None<T>` in the arm `B`.

```typescript
const v: Option<T> = ...;
if (v.isSome) {
  // In this arm A, `v` is `Some<T>` by the type guard method
}
else {
  // In this arm B, `v` should be `None<T>`  by the type guard method,
  // but the current `Option<T>` is just a interface, which cannot restrict type "range".
}
```

So this change make this library more Rust's enum style!

## Details

- Define `type Option<T> = Some<T> | None<T>`.
  - just as a union type.
- Rename `interface Option<T>` to `interface OptionMethods<T>`

## Drawback

####  Make `OptionMethods<T>` non-public interface (BREAKING CHANGE)

- User cannot implement `OptionMethods<T>` to their types.
- But I don't think this would be a serious problem.
   - Who want to implement a self-defined `Option<T>`?
     - They would extend `Some` or `None`.
   - If someone want to implements custom type, we'll republish `OptionMethods<T>`

## Unresolved Questions

none.

## compile perf

I think we don't have serious regression on MacBook Pro 15 (Mid 2014, 2.8GHz Intel Core i7, 16GB 1600MHz DDR3 RAM)


|             |  before   | after    |
|-------------|-----------|----------|
|Files:       |        5  |       5  |
|Lines:       |    19312  |   19314  |
|Nodes:       |    83997  |   84024  |
|Identifiers: |    32589  |   32601  |
|Symbols:     |   116894  |  117648  |
|Types:       |     5969  |    6899  |
|Memory used: |   62181K  |  67498K  |
|I/O read:    |    0.00s  |   0.00s  |
|I/O write:   |    0.00s  |   0.00s  |
|Parse time:  |    0.15s  |   0.15s  |
|Bind time:   |    0.10s  |   0.10s  |
|Check time:  |    0.48s  |   0.52s  |
|Emit time:   |    0.04s  |   0.03s  |
|Total time:  |    0.77s  |   0.80s  |



<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.svg" height="40" alt="Review on Reviewable"/>](https://reviewable.io/reviews/saneyuki/option-t.js/99)
<!-- Reviewable:end -->